도메인 기준 패키지 구조 변경 및 불필요한 파일 삭제
