+++ src/main/java/com/example/voucher/customer/service/dto/CustomerDTO.java
@@ -1,0 +1,9 @@
+package com.example.voucher.customer.service.dto;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+import com.example.voucher.constant.CustomerType;
+
+public record CustomerDTO(UUID customerId, String name, String email, CustomerType customerType,
+                          LocalDateTime createdAt) {
+}
