--- src/main/java/com/example/voucher/domain/wallet/Wallet.java
@@ -1,34 +1,0 @@
-package com.example.voucher.domain.wallet;
-
-import java.util.UUID;
-
-public class Wallet {
-
-    private final UUID walletId;
-    private final UUID customerId;
-    private final UUID voucherId;
-
-    public Wallet(UUID customerId, UUID voucherId) {
-        this.walletId = UUID.randomUUID();
-        this.customerId = customerId;
-        this.voucherId = voucherId;
-    }
-
-    public Wallet(UUID walletId, UUID customerId, UUID voucherId) {
-        this.walletId = walletId;
-        this.customerId = customerId;
-        this.voucherId = voucherId;
-    }
-
-    public UUID getWalletId() {
-        return walletId;
-    }
-
-    public UUID getCustomerId() {
-        return customerId;
-    }
-
-    public UUID getVoucherId() {
-        return voucherId;
-    }
-}
