--- src/main/java/com/example/voucher/service/customer/dto/CustomerDTO.java
@@ -1,9 +1,0 @@
-package com.example.voucher.service.customer.dto;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-import com.example.voucher.constant.CustomerType;
-
-public record CustomerDTO(UUID customerId, String name, String email, CustomerType customerType,
-                          LocalDateTime createdAt) {
-}
