--- src/main/java/com/example/voucher/service/wallet/WalletDTO.java
@@ -1,6 +1,0 @@
-package com.example.voucher.service.wallet;
-
-import java.util.UUID;
-
-public record WalletDTO(UUID walletId, UUID customerId, UUID voucherId) {
-}
