+++ src/main/java/com/example/voucher/voucher/model/Voucher.java
@@ -1,0 +1,23 @@
+package com.example.voucher.voucher.model;
+
+import static com.example.voucher.constant.ExceptionMessage.*;
+import java.util.UUID;
+import com.example.voucher.constant.VoucherType;
+
+public interface Voucher {
+
+    UUID getVoucherId();
+
+    Long getValue();
+
+    VoucherType getVoucherType();
+
+    long discount(long originalAmount);
+
+    default void validatePositive(long value) {
+        if (value <= 0) {
+            throw new IllegalArgumentException(MESSAGE_ERROR_POSITIVE_CONSTRAINT);
+        }
+    }
+
+}
