+++ src/main/java/com/example/voucher/wallet/service/dto/WalletDTO.java
@@ -1,0 +1,6 @@
+package com.example.voucher.wallet.service.dto;
+
+import java.util.UUID;
+
+public record WalletDTO(UUID walletId, UUID customerId, UUID voucherId) {
+}
