+++ src/test/java/com/example/voucher/repository/JdbcVoucherRepositoryTest.java
@@ -12,10 +12,10 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.test.context.ContextConfiguration;
 import com.example.voucher.constant.VoucherType;
-import com.example.voucher.domain.voucher.FixedAmountVoucher;
-import com.example.voucher.domain.voucher.PercentDiscountVoucher;
-import com.example.voucher.domain.voucher.Voucher;
-import com.example.voucher.repository.voucher.JdbcVoucherRepository;
+import com.example.voucher.voucher.model.FixedAmountVoucher;
+import com.example.voucher.voucher.model.PercentDiscountVoucher;
+import com.example.voucher.voucher.model.Voucher;
+import com.example.voucher.voucher.repository.JdbcVoucherRepository;
 
 @JdbcTest
 @ContextConfiguration(classes = JdbcVoucherRepository.class)
